Fixed the list editing, now works with both keyboard and mouse
Editing the list with the keyboard shouldn't behave correctly now. Index
updates as it should in reaction to keyboard presses and the correct
actions that correspond to the buttons pressed should be much more
consistent now.

Made the panel listen for mouse clicks too, so users can now also click
directly on the item in the list they would like to edit.